requirements.txt
discord.py
feedparser
python-dotenv
requests
beautifulsoup4